Create greedy CodesProviderFactory implementation.
This allows the mastermind program to compute results like the original
draft did. The algorithm works by comparing all the guesses it could
make to all the answers that remain. It then sorts each answer into a
bin based on the Response of that comparison. Every guess is assigned a
weight equivalent to the size of the largest bin thereby created. The
guess with the smallest weight is finally chosen as the next and only
guess to make.

When this factory is used, results are cheap to calculate but not
necessarily correct, hence 'greedy'.